Disabled update_marker_set method.
This method was attempting to add additional markers to the OpenSim Model without success. I managed to get it to add a number of these markers, but many of them appear to be duplicates of markers already included in the model. I have disabled this method for now, but if we want to add any additional LowerLimb markers to the model we should use this method to do so.